Refactor some video demo components into classes, seperate files
